Updates to access control
